TIP-221: Move ProductTemplateApplier
